Fixed more bugs and simplified code
Now tasks are oonly executed when all the tasks are retrieved to fix the
problem where the evaluation is ended before all the tasks are
retrieved.

Files are now sent using flask and request. No need to convert to bytes
and send it manually :^).

Removed some redundancy by creating other functions.

Added sleeps to test_runner to fix the error where a node doesnt send
heartbeats for too long.

-.-